pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "mmm-analytics"
version = "1.0.0"
description = "Production-grade Marketing Mix Modeling (MMM) for Ad Tech & Media Analytics"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Sai Santhosh V", email = "saisanthosh@example.com" }
]
maintainers = [
    { name = "Sai Santhosh V", email = "saisanthosh@example.com" }
]
keywords = [
    "marketing-mix-modeling",
    "mmm",
    "adtech",
    "media-analytics",
    "attribution",
    "adstock",
    "saturation",
    "ridge-regression",
    "budget-optimization",
    "roas",
    "marketing-analytics"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0,<3.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "scipy>=1.11.0,<2.0.0",
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.13.0,<1.0.0",
    "click>=8.1.0,<9.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "rich>=13.0.0,<14.0.0",
    "structlog>=24.0.0,<25.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "pytest-xdist>=3.5.0,<4.0.0",
    "hypothesis>=6.92.0,<7.0.0",
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pandas-stubs>=2.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "bandit>=1.7.0,<2.0.0",
    "safety>=2.3.0,<4.0.0"
]
docs = [
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.5.0,<10.0.0",
    "mkdocstrings[python]>=0.24.0,<1.0.0"
]
all = ["mmm-analytics[dev,docs]"]

[project.scripts]
mmm = "mmm_analytics.cli:main"

[project.urls]
Homepage = "https://github.com/saisanthoshv/mmm-analytics"
Documentation = "https://github.com/saisanthoshv/mmm-analytics#readme"
Repository = "https://github.com/saisanthoshv/mmm-analytics.git"
Issues = "https://github.com/saisanthoshv/mmm-analytics/issues"
Changelog = "https://github.com/saisanthoshv/mmm-analytics/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/mmm_analytics",
    "/tests",
    "/configs"
]

[tool.hatch.build.targets.wheel]
packages = ["mmm_analytics"]

# Ruff configuration (linting + formatting)
[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info"
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented out code)
    "RUF",    # Ruff-specific rules
    "NPY",    # NumPy-specific rules
    "PD",     # pandas-vet
    "D",      # pydocstyle
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ARG"]
"scripts/*" = ["D"]

[tool.ruff.lint.isort]
known-first-party = ["mmm_analytics"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

# MyPy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*"
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p", "no:warnings"
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow-running tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

# Coverage configuration
[tool.coverage.run]
source = ["mmm_analytics"]
branch = true
parallel = true
omit = [
    "tests/*",
    "*/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod"
]
fail_under = 80
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Skip assert warnings in tests
